+++ demos/appyx-interactions/android/src/main/kotlin/com/bumble/appyx/interactions/sample/SpotlightExperiment.kt
@@ -16,6 +16,7 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.dp
@@ -128,6 +129,7 @@ fun <InteractionTarget : Any> SpotlightUi(
     color: Color = Color.Unspecified
 ) {
     Children(
+        clipToBounds = false,
         interactionModel = spotlight,
         modifier = modifier
             .padding(
